move editors to own composable, reset errors in addAttendee dialog
