+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/presentation/event_screen/EventScreen.kt
@@ -4,8 +4,10 @@ import android.content.res.Configuration
 import androidx.compose.foundation.*
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.*
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.ChevronRight
@@ -22,11 +24,9 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight.Companion.Bold
 import androidx.compose.ui.text.font.FontWeight.Companion.SemiBold
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.DialogProperties
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
@@ -34,9 +34,7 @@ import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
 import com.realityexpander.tasky.R
 import com.realityexpander.tasky.agenda_feature.domain.AgendaItem
 import com.realityexpander.tasky.agenda_feature.domain.Attendee
-import com.realityexpander.tasky.agenda_feature.presentation.common.components.EditTextModal
 import com.realityexpander.tasky.agenda_feature.presentation.common.components.TimeDateRow
-import com.realityexpander.tasky.agenda_feature.presentation.common.util.toZonedDateTime
 import com.realityexpander.tasky.agenda_feature.presentation.event_screen.EventScreenEvent.*
 import com.realityexpander.tasky.agenda_feature.presentation.event_screen.components.AttendeeList
 import com.realityexpander.tasky.agenda_feature.presentation.event_screen.components.PillButton
@@ -46,11 +44,6 @@ import com.realityexpander.tasky.auth_feature.domain.AuthInfo
 import com.realityexpander.tasky.core.presentation.common.modifiers.*
 import com.realityexpander.tasky.core.presentation.theme.TaskyLightGreen
 import com.realityexpander.tasky.core.presentation.theme.TaskyTheme
-import com.vanpra.composematerialdialogs.MaterialDialog
-import com.vanpra.composematerialdialogs.datetime.date.datepicker
-import com.vanpra.composematerialdialogs.datetime.time.timepicker
-import com.vanpra.composematerialdialogs.rememberMaterialDialogState
-import java.time.LocalDateTime
 import java.time.ZonedDateTime
 import java.util.*
 
@@ -651,211 +644,12 @@ fun AddEventScreenContent(
     }
 
     // • EDITORS FOR EVENT PROPERTIES
-    state.editMode?.let { editMode ->
-        when (editMode) {
-            is EditMode.ChooseTitleText,
-            is EditMode.ChooseDescriptionText -> {
-                editMode as EditMode.TextPayload
-
-                EditTextModal(
-                    text = editMode.text,
-                    title = editMode.dialogTitle.get,
-                    editTextStyle =
-                    if (editMode is EditMode.ChooseTitleText)
-                        MaterialTheme.typography.h2  // cant access from non-compose function, make a wrapper?
-                    else
-                        MaterialTheme.typography.body1, // cant access from non-compose function, make a wrapper?
-                    onSave = {
-                        onAction(EditMode.UpdateText(it))
-                    },
-                    onCancel = {
-                        onAction(CancelEditMode)
-                    }
-                )
-            }
-            is EditMode.ChooseFromDate,
-            is EditMode.ChooseToDate -> {
-                editMode as EditMode.DateTimePayload
-                var pickedDate by remember { mutableStateOf(LocalDateTime.now()) }
-                val dateDialogState = rememberMaterialDialogState()
-
-                dateDialogState.show()
-                MaterialDialog(
-                    dialogState = dateDialogState,
-                    buttons = {
-                        positiveButton(text = stringResource(R.string.ok)) {
-                            onAction(EditMode.UpdateDateTime(pickedDate?.toZonedDateTime()!!))
-                        }
-                        negativeButton(text = stringResource(R.string.cancel)) {
-                            dateDialogState.hide()
-                            onAction(CancelEditMode)
-                        }
-                    }
-                ) {
-                    datepicker(
-                        initialDate = editMode.dateTime.toLocalDate(),
-                        title = editMode.dialogTitle.get,
-                    ) {
-                        pickedDate = it.atTime(editMode.dateTime.toLocalTime())
-                    }
-                }
-            }
-            is EditMode.ChooseFromTime,
-            is EditMode.ChooseToTime -> {
-                editMode as EditMode.DateTimePayload
-                var pickedTime by remember { mutableStateOf(LocalDateTime.now()) }
-                val dateDialogState = rememberMaterialDialogState()
-
-                dateDialogState.show()
-                MaterialDialog(
-                    dialogState = dateDialogState,
-                    buttons = {
-                        positiveButton(text = stringResource(R.string.ok)) {
-                            onAction(EditMode.UpdateDateTime(pickedTime?.toZonedDateTime()!!))
-                        }
-                        negativeButton(text = stringResource(R.string.cancel)) {
-                            dateDialogState.hide()
-                            onAction(CancelEditMode)
-                        }
-                    }
-                ) {
-                    timepicker(
-                        initialTime = editMode.dateTime.toLocalTime(),
-                        title = editMode.dialogTitle.get,
-                    ) {
-                        pickedTime = it.atDate(editMode.dateTime.toLocalDate())
-                    }
-                }
-            }
-            is EditMode.ChooseRemindAtDateTime -> { // handled in the RemindAt UI element, this is here to remove compiler warning
-            }
-            is EditMode.ChooseAddPhoto -> TODO()
-            is EditMode.ConfirmRemovePhoto -> TODO()
-            is EditMode.ChooseAddAttendee -> {
-                val addAttendeeDialogState = rememberMaterialDialogState()
-                var attendeeEmail by remember { mutableStateOf("") }
-
-                addAttendeeDialogState.show()
-                MaterialDialog (
-                    dialogState = addAttendeeDialogState,
-                    properties = DialogProperties(
-                        dismissOnBackPress = true,
-                        dismissOnClickOutside = true
-                    ),
-                    buttons = {
-                        positiveButton(text = stringResource(R.string.ok)) {
-                            onAction(ValidateAttendeeEmailExistsThenAddAttendee(attendeeEmail))
-                        }
-                        negativeButton(text = stringResource(R.string.cancel)) {
-                            addAttendeeDialogState.hide()
-                            onAction(CancelEditMode)
-                        }
-                    },
-                    onCloseRequest = {
-                        addAttendeeDialogState.hide()
-                        onAction(CancelEditMode)
-                    }
-                ) {
-                    Spacer(modifier = Modifier.tinyHeight())
-
-                    Text(
-                        text = stringResource(R.string.add_attendee_dialog_title),
-                        style = MaterialTheme.typography.h4,
-                        color = MaterialTheme.colors.onSurface,
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                    )
-                    Spacer(modifier = Modifier.mediumHeight())
-
-                    OutlinedTextField(
-                        value = attendeeEmail,
-                        onValueChange = {
-                            attendeeEmail = it
-                            onAction(ClearAddAttendeeDialogErrorMessage)
-                        },
-                        label = { Text(stringResource(R.string.add_attendee_dialog_email_title)) },
-                        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email),
-                        trailingIcon = {
-                            if (state.isProgressVisible) {
-                                CircularProgressIndicator(
-                                    color = MaterialTheme.colors.primary,
-                                    modifier = Modifier.size(24.dp)
-                                )
-                            }
-                        },
-                        modifier = Modifier.fillMaxWidth()
-                            .padding(start = DP.small, end = DP.small)
-                    )
-                    Spacer(modifier = Modifier.smallHeight())
-
-                    // • ERROR MESSAGE
-                    Text(
-                        text = state.addAttendeeDialogErrorMessage?.get ?: "",
-                        style = MaterialTheme.typography.body1,
-                        color = MaterialTheme.colors.error,
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                    )
-                    Spacer(modifier = Modifier.smallHeight())
-                }
-            }
-            is EditMode.ConfirmRemoveAttendee -> {
-                val attendee = editMode.attendee
-                val deleteAttendeeDialogState = rememberMaterialDialogState()
-                onAction(SetIsEditable(true  )) // turn on edit mode when removing attendee
-
-                deleteAttendeeDialogState.show()
-                MaterialDialog (
-                    dialogState = deleteAttendeeDialogState,
-                    properties = DialogProperties(
-                        dismissOnBackPress = true,
-                        dismissOnClickOutside = true
-                    ),
-                    buttons = {
-                        positiveButton(text = stringResource(R.string.remove_attendee_dialog_remove_button)) {
-                            onAction(EditMode.RemoveAttendee(attendee.id))
-                        }
-                        negativeButton(text = stringResource(R.string.cancel)) {
-                            deleteAttendeeDialogState.hide()
-                            onAction(CancelEditMode)
-                        }
-                    },
-                    onCloseRequest = {
-                        deleteAttendeeDialogState.hide()
-                        onAction(CancelEditMode)
-                    }
-                ) {
-                    Spacer(modifier = Modifier.tinyHeight())
-
-                    Text(
-                        text = stringResource(R.string.event_dialog_title_remove_attendee),
-                        style = MaterialTheme.typography.h4,
-                        color = MaterialTheme.colors.onSurface,
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                    )
-                    Spacer(modifier = Modifier.mediumHeight())
-
-                    Text(
-                        text = attendee.fullName,
-                        textAlign = TextAlign.Center,
-                        style = MaterialTheme.typography.h4,
-                        modifier = Modifier.fillMaxWidth()
-                            .padding(start = DP.small, end = DP.small)
-                    )
-                    Text(
-                        text = attendee.email,
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier.fillMaxWidth()
-                            .padding(start = DP.small, end = DP.small)
-                    )
-                    Spacer(modifier = Modifier.smallHeight())
-                }
-            }
-        }
+    state.editMode?.let {
+        Editors(
+            editMode = it,
+            state = state,
+            onAction = onAction
+        )
     }
 }
