+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/presentation/event_screen/EventScreenState.kt
@@ -17,6 +17,7 @@ data class EventScreenState(
     val errorMessage: UiText? = null,
 
     val addAttendeeDialogErrorMessage: UiText? = null,
+    val isAttendeeEmailValid: Boolean? = null,
 
     // Stateful One-time events
 //    val scrollToItemId: UuidStr? = null,  // todo add one-time events
